create train and dev files instead of just training files
